requirements.txt
torchaudio==2.8.0
soundfile
numpy<2.0
speechbrain==1.0.0
torch
fastapi
uvicorn
python-multipart
pyaudio
onnx 
onnxruntime-gpu==1.16.3
